Updates gpm script to support cloning private repos
Go's 'go get' functionality lacks support for cloning private
repositories via SSH.  This commit extends the core 'gpm' script to
bootstrap cloning of private repositories.

The core change is to suppot an additional '.git' suffix for import
paths in the Godeps file.  If this suffix is present AND the package
does not already exist in GOPATH, 'git clone' will be run to pull down
the initial copy of the repo (instead of the standard 'go get -u -d').
Subsequent invocations of gpm will still use the defined checkout
command for each VCS gpm supports.